+++ packages/moxb/src/table/TableColumnImpl.ts
@@ -1,73 +1,37 @@
 import { action, observable } from 'mobx';
 import { Table } from './Table';
-import { SortDirection, TableColumn } from './TableColumn';
+import { TableColumn } from './TableColumn';
+import { SortDirection } from './TableSort';
+import { BindImpl, BindOptions } from '../bind/BindImpl';
 
-export interface TableColumnOptions {
-    header: string;
-    accessor?: string;
+export interface TableColumnOptions extends BindOptions {
     sortable?: boolean;
-    isInitialSort?: boolean;
+    preferredSortDirection?: SortDirection;
     width?: number;
 }
 
-export class TableColumnImpl implements TableColumn {
-    private readonly impl: TableColumnOptions;
-    private readonly table: Table<any>;
-    @observable sortDirection: SortDirection;
-
-    constructor(impl: TableColumnOptions, tableImpl: Table<any>) {
-        this.impl = impl;
-        this.table = tableImpl;
-        // Set the initial sort filter column
-        this.sortDirection = this.impl.isInitialSort ? 'descending' : undefined;
-
-        if (this.impl.sortable) {
-            if (!this.impl.accessor) {
-                console.error(`To make a column sortable, you also have to set the accessor property.`);
-            }
-        }
+export class TableColumnImpl extends BindImpl<TableColumnOptions> implements TableColumn {
+    readonly column: string;
+    constructor(impl: TableColumnOptions, private readonly table: Table<any>) {
+        super({ ...impl, id: table.id + '.' + impl.id });
+        this.column = impl.id;
     }
 
     @action.bound
-    onClick(): void {
-        if (this.table.sortAccessor !== this.accessor) {
-            // Reset the UI props before setting the newly selected column
-            this.table.columns!.forEach((col: TableColumn) => col.setSortDirection(undefined));
-            this.table.setSortAccessor(this.accessor);
-            this.table.setSortDirection('descending');
-        } else {
-            this.table.setSortDirection(this.sortDirection === 'ascending' ? 'descending' : 'ascending');
-        }
-        this.sortDirection = this.table.sortDirection;
+    toggleSort() {
+        this.table.sort.toggleSort(this.column, this.preferredSortDirection);
     }
 
-    @action.bound
-    setSortDirection(sortDirection: SortDirection): void {
-        this.sortDirection = sortDirection;
+    get preferredSortDirection() {
+        return this.impl.preferredSortDirection;
     }
 
-    get header() {
-        if (this.impl.header) {
-            return this.impl.header;
+    get sortDirection() {
+        const sort = this.table.sort.sort[0];
+        if (sort && sort.column === this.column) {
+            return sort.sortDirection;
         }
-    }
-
-    get accessor() {
-        if (this.impl.accessor) {
-            return this.impl.accessor;
-        }
-    }
-
-    get sortable() {
-        if (this.impl.sortable) {
-            return this.impl.sortable;
-        }
-    }
-
-    get isInitialSort() {
-        if (this.impl.isInitialSort) {
-            return this.impl.isInitialSort;
-        }
+        return undefined;
     }
 
     get width() {