+++ packages/moxb/src/table/__tests__/TableColumnImpl.test.ts
@@ -1,4 +1,3 @@
-import * as moxb from '../../index';
 import { Table } from '../Table';
 import { TableColumn } from '../TableColumn';
 import { TableColumnImpl } from '../TableColumnImpl';
@@ -9,20 +8,22 @@ describe('isInitialSort', function() {
 
     beforeEach(function() {
         bindTable = new TableImpl({
-            columns: bind => [
-                new moxb.TableColumnImpl({ header: 'test', accessor: 'test', isInitialSort: true }, bind),
-                new moxb.TableColumnImpl({ header: 'test2', accessor: 'test2' }, bind),
+            id: 'table',
+            columns: table => [
+                new TableColumnImpl({ id: 'test', label: 'Test' }, table),
+                new TableColumnImpl({ id: 'test2', label: 'Test 2' }, table),
             ],
+            data: () => [],
+            initialSort: [{ column: 'test2', sortDirection: 'descending' }],
         });
     });
 
     it('should set the initial sort column', function() {
-        expect(bindTable.columns![0].sortDirection).not.toBeUndefined();
-        expect(bindTable.columns![0].sortDirection).toBe('descending');
+        expect(bindTable.columns[1].sortDirection).toBe('descending');
     });
 
     it('should set the column sorted to undefined, if isInitialSort value was not set', function() {
-        expect(bindTable.columns![1].sortDirection).toBeUndefined();
+        expect(bindTable.columns[0].sortDirection).toBeUndefined();
     });
 });
 
@@ -36,67 +37,37 @@ describe('sortable', function() {
         onClick = jest.fn().mockImplementation(function(this: any) {
             thisClick = this;
         });
-        bindTable = new TableImpl({ columns: () => [] });
+        bindTable = new TableImpl({
+            id: 'table',
+            columns: table => [
+                new TableColumnImpl({ id: 'col1' }, table),
+                new TableColumnImpl({ id: 'col.asc', preferredSortDirection: 'ascending' }, table),
+                new TableColumnImpl({ id: 'col.desc', preferredSortDirection: 'descending' }, table),
+            ],
+            data: () => [],
+        });
     });
 
     it('should not make the column clickable for missing sorting prop', function() {
-        const column: TableColumn = new moxb.TableColumnImpl(
+        const column: TableColumn = new TableColumnImpl(
             {
-                header: 'test',
-                accessor: 'test',
+                id: 'test',
             },
             bindTable
         );
-        const columnOnClick = jest.spyOn(column, 'onClick');
+        const columnOnClick = jest.spyOn(column, 'toggleSort');
         columnOnClick.mockImplementation(onClick);
-        column.onClick();
+        column.toggleSort();
         expect(thisClick).not.toBeUndefined();
-        expect(column.onClick).toBeCalled();
+        expect(column.toggleSort).toBeCalled();
     });
 
     it('should make the column clickable', function() {
-        const column: TableColumn = new moxb.TableColumnImpl(
-            {
-                header: 'test',
-                accessor: 'test',
-                sortable: true,
-            },
-            bindTable
-        );
-        const columnOnClick = jest.spyOn(column, 'onClick');
+        const column = bindTable.columns[0];
+        const columnOnClick = jest.spyOn(column, 'toggleSort');
         columnOnClick.mockImplementation(onClick);
-        column.onClick();
+        column.toggleSort();
         expect(thisClick).not.toBeUndefined();
-        expect(column.onClick).toBeCalled();
-    });
-
-    it('should call console.warn, if the accessor prop was not set', function() {
-        const consoleErr = jest.spyOn(console, 'error');
-        consoleErr.mockImplementation(() => {});
-        const column: TableColumn = new moxb.TableColumnImpl(
-            {
-                header: 'test',
-                sortable: true,
-            },
-            bindTable
-        );
-        expect(column).toBeInstanceOf(TableColumnImpl);
-        expect(consoleErr).toHaveBeenCalled();
-        consoleErr.mockRestore();
-    });
-
-    it('should call console.warn, if the onClick function was not set', function() {
-        const consoleErr = jest.spyOn(console, 'error');
-        consoleErr.mockImplementation(() => {});
-        const column: TableColumn = new moxb.TableColumnImpl(
-            {
-                header: 'test',
-                sortable: true,
-            },
-            bindTable
-        );
-        expect(column).toBeInstanceOf(TableColumnImpl);
-        expect(consoleErr).toHaveBeenCalled();
-        consoleErr.mockRestore();
+        expect(column.toggleSort).toBeCalled();
     });
 });